Refactor Navigator
