+++ src/libs/Navigation/AppNavigator/DrawerNavigator/index.native.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import withWindowDimensions from '../../../../components/withWindowDimensions';
+import BaseDrawerNavigator from './BaseDrawerNavigator';
+import {propTypes, defaultProps} from './drawerNavigatorPropTypes';
+
+const DrawerNavigator = props => (
+    <BaseDrawerNavigator
+        // eslint-disable-next-line react/jsx-props-no-spreading
+        {...props}
+        useLegacyImplementation={false}
+    />
+);
+
+DrawerNavigator.propTypes = propTypes;
+DrawerNavigator.defaultProps = defaultProps;
+DrawerNavigator.displayName = 'DrawerNavigator';
+export default withWindowDimensions(DrawerNavigator);
