+++ src/libs/Navigation/AppNavigator/MainDrawerNavigator.js
@@ -11,7 +11,7 @@ import Permissions from '../../Permissions';
 // Screens
 import ReportScreen from '../../../pages/home/ReportScreen';
 import SidebarScreen from '../../../pages/home/sidebar/SidebarScreen';
-import BaseDrawerNavigator from './BaseDrawerNavigator';
+import DrawerNavigator from './DrawerNavigator';
 import {findLastAccessedReport} from '../../reportUtils';
 
 const propTypes = {
@@ -56,7 +56,7 @@ const MainDrawerNavigator = (props) => {
     // This way routing information is updated (if needed) based on the initial report ID resolved.
     // This is usually needed after login/create account and re-launches
     return (
-        <BaseDrawerNavigator
+        <DrawerNavigator
             drawerContent={() => <SidebarScreen />}
             screens={[
                 {
